pc99, SerialServer: restructure to allow for other device types besides COM ports
