+++ components/SerialServer/src/plat/pc99/com/plat.c
@@ -11,8 +11,8 @@
 #include <platsupport/irq.h>
 #include <sel4utils/sel4_zf_logif.h>
 
-#include "../../plat.h"
-#include "../../serial.h"
+#include "../../../plat.h"
+#include "../../../serial.h"
 
 void plat_post_init(ps_irq_ops_t *irq_ops)
 {
